+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Visit the package at [pub.dev](https://pub.dev/packages/solidpod).
 
 ## 0.10 Further UI migrations
 
-+ Allow sharing externally owned files [0.9.5 20260110 jesscmoore]
++ Allow sharing externally owned files [0.9.6 20260110 jesscmoore]
++ Specify owner and granter webid when sharing files [0.9.5 20260112 jesscmoore]
 + Fixes individual recipient dialog background [0.9.4 20260113 jesscmoore]
 + Fix sharing page heading colours [0.9.3 20260113 jesscmoore]
 + Fixes key creation for encrypting files [0.9.2 20260112 dc]